backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-docx
sqlalchemy
psycopg2-binary
pydantic
